amp: Move to amp
