+++ internal/service/amp/alert_manager_definition.go
@@ -1,4 +1,4 @@
-package prometheus
+package amp
 
 import (
 	"context"
