+++ internal/service/amp/rule_group_namespace.go
@@ -1,4 +1,4 @@
-package prometheus
+package amp
 
 import (
 	"context"
