+++ internal/service/amp/status.go
@@ -1,4 +1,4 @@
-package prometheus
+package amp
 
 import (
 	"context"
