+++ internal/service/amp/wait.go
@@ -1,4 +1,4 @@
-package prometheus
+package amp
 
 import (
 	"context"
